+++ frontend/src/views/common/Login.vue
@@ -15,7 +15,7 @@ const props = defineProps({
     bindUserAddress: {
         type: Function,
         default: async () => { await api.bindUserAddress(); },
-        requried: true
+        required: true
     },
     newAddressPath: {
         type: Function,
@@ -29,7 +29,7 @@ const props = defineProps({
                 }),
             });
         },
-        requried: true
+        required: true
     },
 })
 
@@ -72,7 +72,7 @@ const { locale, t } = useI18n({
             login: 'Login',
             pleaseGetNewEmail: 'Please login or click "Get New Email" button to get a new email address',
             getNewEmail: 'Create New Email',
-            getNewEmailTip1: 'Please input the email you want to use. only allow a-z and 0-9',
+            getNewEmailTip1: 'Please input the email you want to use. only allow: ',
             getNewEmailTip2: 'Levaing it blank will generate a random email address.',
             getNewEmailTip3: 'You can choose a domain from the dropdown list.',
             credential: 'Email Address Credential',
@@ -87,7 +87,7 @@ const { locale, t } = useI18n({
             login: '登录',
             pleaseGetNewEmail: '请"登录"或点击 "注册新邮箱" 按钮来获取一个新的邮箱地址',
             getNewEmail: '创建新邮箱',
-            getNewEmailTip1: '请输入你想要使用的邮箱地址, 只允许 a-z, 0-9',
+            getNewEmailTip1: '请输入你想要使用的邮箱地址, 只允许: ',
             getNewEmailTip2: '留空将会生成一个随机的邮箱地址。',
             getNewEmailTip3: '你可以从下拉列表中选择一个域名。',
             credential: '邮箱地址凭据',
@@ -101,6 +101,18 @@ const { locale, t } = useI18n({
     }
 });
 
+const addressRegex = computed(() => {
+    try {
+        if (openSettings.value.addressRegex) {
+            return new RegExp(openSettings.value.addressRegex, 'g');
+        }
+    } catch (error) {
+        console.error(error);
+        message.error(`Invalid addressRegex: ${openSettings.value.addressRegex}`);
+    }
+    return /[^a-z0-9]/g;
+});
+
 const generateNameLoading = ref(false);
 const generateName = async () => {
     try {
@@ -110,7 +122,7 @@ const generateName = async () => {
             .split('@')[0]
             .replace(/\s+/g, '.')
             .replace(/\.{2,}/g, '.')
-            .replace(/[^a-z0-9]/g, '')
+            .replace(addressRegex.value, '')
             .toLowerCase();
     } catch (error) {
         message.error(error.message || "error");
@@ -181,7 +193,7 @@ onMounted(async () => {
         <n-alert v-if="userSettings.user_email" :show-icon="false" :bordered="false" closable>
             <span>{{ t('bindUserInfo') }}</span>
         </n-alert>
-        <n-tabs v-model:value="tabValue" size="large" justify-content="space-evenly">
+        <n-tabs v-if="openSettings.fetched" v-model:value="tabValue" size="large" justify-content="space-evenly">
             <n-tab-pane name="signin" :tab="t('login')">
                 <n-form>
                     <n-form-item-row :label="t('credential')" required>
@@ -206,7 +218,7 @@ onMounted(async () => {
                 <n-spin :show="generateNameLoading">
                     <n-form>
                         <span>
-                            <p>{{ t("getNewEmailTip1") }}</p>
+                            <p>{{ t("getNewEmailTip1") + addressRegex.source }}</p>
                             <p>{{ t("getNewEmailTip2") }}</p>
                             <p>{{ t("getNewEmailTip3") }}</p>
                         </span>