+++ worker/src/admin_api/index.ts
@@ -8,6 +8,9 @@ import { CONSTANTS } from '../constants'
 import cleanup_api from './cleanup_api'
 import admin_user_api from './admin_user_api'
 import webhook_settings from './webhook_settings'
+import mail_webhook_settings from './mail_webhook_settings'
+import oauth2_settings from './oauth2_settings'
+import worker_config from './worker_config'
 
 export const api = new Hono<HonoCustomType>()
 
@@ -40,7 +43,13 @@ api.post('/admin/new_address', async (c) => {
         return c.text("Please provide a name", 400)
     }
     try {
-        const res = await newAddress(c, name, domain, enablePrefix, false, null, false);
+        const res = await newAddress(c, {
+            name, domain, enablePrefix,
+            checkLengthByConfig: false,
+            addressPrefix: null,
+            checkAllowDomains: false,
+            enableCheckNameRegex: false,
+        });
         return c.json(res);
     } catch (e) {
         return c.text(`Failed create address: ${(e as Error).message}`, 400)
@@ -246,10 +255,12 @@ api.get('/admin/account_settings', async (c) => {
         const blockList = await getJsonSetting(c, CONSTANTS.ADDRESS_BLOCK_LIST_KEY);
         const sendBlockList = await getJsonSetting(c, CONSTANTS.SEND_BLOCK_LIST_KEY);
         const verifiedAddressList = await getJsonSetting(c, CONSTANTS.VERIFIED_ADDRESS_LIST_KEY);
+        const fromBlockList = c.env.KV ? await c.env.KV.get<string[]>(CONSTANTS.EMAIL_KV_BLACK_LIST, 'json') : [];
         return c.json({
             blockList: blockList || [],
             sendBlockList: sendBlockList || [],
-            verifiedAddressList: verifiedAddressList || []
+            verifiedAddressList: verifiedAddressList || [],
+            fromBlockList: fromBlockList || []
         })
     } catch (error) {
         console.error(error);
@@ -259,7 +270,7 @@ api.get('/admin/account_settings', async (c) => {
 
 api.post('/admin/account_settings', async (c) => {
     /** @type {{ blockList: Array<string>, sendBlockList: Array<string> }} */
-    const { blockList, sendBlockList, verifiedAddressList } = await c.req.json();
+    const { blockList, sendBlockList, verifiedAddressList, fromBlockList } = await c.req.json();
     if (!blockList || !sendBlockList || !verifiedAddressList) {
         return c.text("Invalid blockList or sendBlockList", 400)
     }
@@ -278,14 +289,23 @@ api.post('/admin/account_settings', async (c) => {
         c, CONSTANTS.VERIFIED_ADDRESS_LIST_KEY,
         JSON.stringify(verifiedAddressList)
     )
+    if (fromBlockList?.length > 0 && !c.env.KV) {
+        return c.text("Please enable KV to use fromBlockList", 400)
+    }
+    if (fromBlockList) {
+        await c.env.KV.put(CONSTANTS.EMAIL_KV_BLACK_LIST, JSON.stringify(fromBlockList || []))
+    }
     return c.json({
         success: true
     })
 })
 
+// cleanup
 api.post('/admin/cleanup', cleanup_api.cleanup)
 api.get('/admin/auto_cleanup', cleanup_api.getCleanup)
 api.post('/admin/auto_cleanup', cleanup_api.saveCleanup)
+
+// user settings
 api.get('/admin/user_settings', admin_user_api.getSetting)
 api.post('/admin/user_settings', admin_user_api.saveSetting)
 api.get('/admin/users', admin_user_api.getUsers)
@@ -294,5 +314,19 @@ api.post('/admin/users', admin_user_api.createUser)
 api.post('/admin/users/:user_id/reset_password', admin_user_api.resetPassword)
 api.get('/admin/user_roles', async (c) => c.json(getUserRoles(c)))
 api.post('/admin/user_roles', admin_user_api.updateUserRoles)
+
+// user oauth2 settings
+api.get('/admin/user_oauth2_settings', oauth2_settings.getUserOauth2Settings)
+api.post('/admin/user_oauth2_settings', oauth2_settings.saveUserOauth2Settings)
+
+// webhook settings
 api.get("/admin/webhook/settings", webhook_settings.getWebhookSettings);
 api.post("/admin/webhook/settings", webhook_settings.saveWebhookSettings);
+
+// mail webhook settings
+api.get("/admin/mail_webhook/settings", mail_webhook_settings.getWebhookSettings);
+api.post("/admin/mail_webhook/settings", mail_webhook_settings.saveWebhookSettings);
+api.post("/admin/mail_webhook/test", mail_webhook_settings.testWebhookSettings);
+
+// worker config
+api.get("/admin/worker/configs", worker_config.getConfig);