+++ worker/src/common.ts
@@ -4,17 +4,66 @@ import { Jwt } from 'hono/utils/jwt'
 import { getBooleanValue, getDomains, getStringValue, getIntValue, getUserRoles, getDefaultDomains } from './utils';
 import { HonoCustomType, UserRole } from './types';
 import { unbindTelegramByAddress } from './telegram_api/common';
+import { CONSTANTS } from './constants';
+import { AdminWebhookSettings, WebhookMail, WebhookSettings } from './models';
+
+const DEFAULT_NAME_REGEX = /[^a-z0-9]/g;
+
+const checkNameRegex = (c: Context<HonoCustomType>, name: string) => {
+    let error = null;
+    try {
+        const regexStr = getStringValue(c.env.ADDRESS_CHECK_REGEX);
+        if (!regexStr) return;
+        const regex = new RegExp(regexStr);
+        if (!regex.test(name)) {
+            error = new Error(`Name not match regex: /${regexStr}/`);
+        }
+    }
+    catch (e) {
+        console.error("Failed to check address regex", e);
+    }
+    if (error) {
+        throw error;
+    }
+}
+
+const getNameRegex = (c: Context<HonoCustomType>): RegExp => {
+    try {
+        const regex = getStringValue(c.env.ADDRESS_REGEX);
+        if (!regex) {
+            return DEFAULT_NAME_REGEX;
+        }
+        return new RegExp(regex, 'g');
+    }
+    catch (e) {
+        console.error("Failed to get address regex", e);
+    }
+    return DEFAULT_NAME_REGEX;
+}
 
 export const newAddress = async (
     c: Context<HonoCustomType>,
-    name: string, domain: string | undefined | null,
-    enablePrefix: boolean,
-    checkLengthByConfig: boolean = true,
-    addressPrefix: string | undefined | null = null,
-    checkAllowDomains: boolean = true
+    {
+        name,
+        domain,
+        enablePrefix,
+        checkLengthByConfig = true,
+        addressPrefix = null,
+        checkAllowDomains = true,
+        enableCheckNameRegex = true,
+    }: {
+        name: string, domain: string | undefined | null,
+        enablePrefix: boolean,
+        checkLengthByConfig?: boolean,
+        addressPrefix?: string | undefined | null,
+        checkAllowDomains?: boolean,
+        enableCheckNameRegex?: boolean,
+    }
 ): Promise<{ address: string, jwt: string }> => {
+    // check name
+    if (enableCheckNameRegex) checkNameRegex(c, name);
     // remove special characters
-    name = name.replace(/[^a-z0-9]/g, '')
+    name = name.replace(getNameRegex(c), '')
     // name min length min 1
     const minAddressLength = Math.max(
         checkLengthByConfig ? getIntValue(c.env.MIN_ADDRESS_LEN, 1) : 1,
@@ -40,6 +89,11 @@ export const newAddress = async (
     }
     // check domain
     const allowDomains = checkAllowDomains ? await getAllowDomains(c) : getDomains(c);
+    // if domain is not set, use the first domain
+    if (!domain && allowDomains.length > 0) {
+        domain = allowDomains[0];
+    }
+    // check domain is valid
     if (!domain || !allowDomains.includes(domain)) {
         throw new Error("Invalid domain")
     }
@@ -78,7 +132,7 @@ export const cleanup = async (
     cleanType: string | undefined | null,
     cleanDays: number | undefined | null
 ): Promise<boolean> => {
-    if (!cleanType || !cleanDays || cleanDays < 0 || cleanDays > 30) {
+    if (!cleanType || typeof cleanDays !== 'number' || cleanDays < 0 || cleanDays > 30) {
         throw new Error("Invalid cleanType or cleanDays")
     }
     console.log(`Cleanup ${cleanType} before ${cleanDays} days`);
@@ -251,4 +305,77 @@ export const getAllowDomains = async (c: Context<HonoCustomType>): Promise<strin
     }
     const user_role = await commonGetUserRole(c, user.user_id);
     return user_role?.domains || getDefaultDomains(c);;
+}
+
+export async function sendWebhook(settings: WebhookSettings, formatMap: WebhookMail): Promise<{ success: boolean, message?: string }> {
+    // send webhook
+    let body = settings.body;
+    for (const key of Object.keys(formatMap)) {
+        /* eslint-disable no-useless-escape */
+        body = body.replace(
+            new RegExp(`\\$\\{${key}\\}`, "g"),
+            JSON.stringify(
+                formatMap[key as keyof WebhookMail]
+            ).replace(/^"(.*)"$/, '\$1')
+        );
+        /* eslint-enable no-useless-escape */
+    }
+    const response = await fetch(settings.url, {
+        method: settings.method,
+        headers: JSON.parse(settings.headers),
+        body: body
+    });
+    if (!response.ok) {
+        console.log("send webhook error", response.status, response.statusText);
+        return { success: false, message: `send webhook error: ${response.status} ${response.statusText}` };
+    }
+    return { success: true }
+}
+
+export async function triggerWebhook(
+    c: Context<HonoCustomType>,
+    address: string,
+    raw_mail: string
+): Promise<void> {
+    if (!c.env.KV || !getBooleanValue(c.env.ENABLE_WEBHOOK)) {
+        return
+    }
+    const webhookList: WebhookSettings[] = []
+
+    // admin mail webhook
+    const adminMailWebhookSettings = await c.env.KV.get<WebhookSettings>(CONSTANTS.WEBHOOK_KV_ADMIN_MAIL_SETTINGS_KEY, "json");
+    if (adminMailWebhookSettings?.enabled) {
+        webhookList.push(adminMailWebhookSettings)
+    }
+
+    // user mail webhook
+    const adminSettings = await c.env.KV.get<AdminWebhookSettings>(CONSTANTS.WEBHOOK_KV_SETTINGS_KEY, "json");
+    if (adminSettings?.allowList.includes(address)) {
+        const settings = await c.env.KV.get<WebhookSettings>(
+            `${CONSTANTS.WEBHOOK_KV_USER_SETTINGS_KEY}:${address}`, "json"
+        );
+        if (settings?.enabled) {
+            webhookList.push(settings)
+        }
+    }
+
+    // no webhook
+    if (webhookList.length === 0) {
+        return
+    }
+    const parsedEmail = await commonParseMail(raw_mail);
+    const webhookMail = {
+        from: parsedEmail?.sender || "",
+        to: address,
+        subject: parsedEmail?.subject || "",
+        raw: raw_mail,
+        parsedText: parsedEmail?.text || "",
+        parsedHtml: parsedEmail?.html || ""
+    }
+    for (const settings of webhookList) {
+        const res = await sendWebhook(settings, webhookMail);
+        if (!res.success) {
+            console.error(res.message);
+        }
+    }
 }